+++ ipaas/src/components/SchemaConfigForm/FormElements/ObjectElement.tsx
@@ -18,7 +18,8 @@
 
 import { Box } from '@wso2/oxygen-ui';
 import type { Dispatch, SetStateAction } from 'react';
-import { type BaseType, type JSONSchema, type LinkingInfo, generateArrayJsonPath, generateMapJsonPath, generateObjectJsonPath, generatedNestedLabel, isBaseType } from '../schemaUtils';
+import type { BaseType, JSONSchema } from '../../../types/schema';
+import { type LinkingInfo, generateArrayJsonPath, generateMapJsonPath, generateObjectJsonPath, generatedNestedLabel, isBaseType } from '../schemaUtils';
 import { AnyOfElement } from './AnyOfElement';
 import { ArrayElement } from './ArrayElement';
 import { BaseElement } from './BaseElement';
