+++ ipaas/src/components/SchemaConfigForm/schemaUtils.ts
@@ -16,35 +16,12 @@
  * under the License.
  */
 
-export interface JSONSchema {
-  $schema?: string;
-  $id?: string;
-  title?: string;
-  name?: string;
-  description?: string;
-  type: 'object' | 'array' | 'string' | 'number' | 'boolean' | 'null' | 'secret';
-  properties?: Record<string, JSONSchema>;
-  required?: string[];
-  requiredLevel?: number;
-  items?: JSONSchema;
-  enum?: string[] | number[];
-  default?: BaseType;
-  format?: string;
-  minLength?: number;
-  maxLength?: number;
-  minimum?: number;
-  maximum?: number;
-  pattern?: string;
-  additionalProperties?: boolean | JSONSchema;
-  anyOf?: JSONSchema[];
-}
+import type { BaseType, JSONSchema } from '../../types/schema';
 
 export interface SchemaAtLevel {
   schema: JSONSchema;
   path: string;
 }
-
-export type BaseType = string | number | boolean;
 
 export type LinkingInfo = { configGroupId?: string; configKeyId?: string; isDynamic?: boolean };
